+++ app/(tabs)/_layout.tsx
@@ -4,7 +4,7 @@ import { Platform } from 'react-native';
 import { HapticTab } from '@/components/premade-comps/HapticTab';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import TabBarBackground from '@/components/ui/TabBarBackground';
-import { useAppContext } from '@/context/AppContext';
+import { useAppContext } from '@/context/AuthContext';
 
 export default function TabLayout() {
