+++ context/ContextProvider.tsx
@@ -1,0 +1,11 @@
+import React, { ReactNode } from "react";
+import { AuthProvider } from "./AuthContext";
+import { UserProvider } from "./UserContext";
+
+export const ContextProvider = ({ children }: { children: ReactNode }) => {
+  return (
+    <AuthProvider>
+      <UserProvider>{children}</UserProvider>
+    </AuthProvider>
+  );
+};
